// Desarrollo_Web_en_entorno_cliente/Casa/Trabajos/Unit1 - JS/foodscore-client/src/interfaces/restaurant.ts
import { User } from "./user";

export interface Restaurant {
    id?: number;
    name: string;
    description: string;
    cuisine: string;
    daysOpen: string[];
    image: string;
    phone: string;
    creator?: User;
    mine?: boolean;
    distance?: number;
    commented?: boolean;
    stars?: number;
    address: string;
    lat: number;
    lng: number;
}